orders/requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
Django==2.2.13
django-filter==2.4.0
djangorestframework==3.12.4
flake8==3.9.2
idna==3.2
mccabe==0.6.1
psycopg2==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2018.5
PyYAML==5.4.1
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.6
